buf: impl size_hint for str types + reorg tests
